Undo import rename
