+++ backend/api/controllers/config.ts
@@ -1,5 +1,5 @@
 import ConfigDAO from "../../dao/config";
-import { MonkeyResponse } from "../../utils/monkey-response";
+import { MonkeyResponse } from "../../handlers/monkey-response";
 
 class ConfigController {
   static async getConfig(req: MonkeyTypes.Request): Promise<MonkeyResponse> {
