+++ backend/api/controllers/preset.ts
@@ -1,5 +1,5 @@
 import PresetDAO from "../../dao/preset";
-import { MonkeyResponse } from "../../utils/monkey-response";
+import { MonkeyResponse } from "../../handlers/monkey-response";
 
 class PresetController {
   static async getPresets(req: MonkeyTypes.Request): Promise<MonkeyResponse> {
