+++ backend/api/routes/users.ts
@@ -4,7 +4,7 @@ import { Router } from "express";
 import UserController from "../controllers/user";
 import { asyncHandler, validateRequest } from "../../middlewares/api-utils";
 import * as RateLimit from "../../middlewares/rate-limit";
-import { isUsernameValid } from "../../utils/validation";
+import { isUsernameValid } from "../../handlers/validation";
 
 const router = Router();
