+++ backend/handlers/auth.js
