+++ backend/handlers/captcha.js
